Update DB, new functionalities

+++ Backend/src/main/java/com/example/demo/ProductRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 public interface ProductRepository extends CrudRepository<Product, Long> {
 
     Product findByName(String name);
-    Product findByPrice(int price);
+    Product findByPrice(long price);
     Product findById(String id);
 }
